Move pool to it's own module
