+++ src/transport/smtp/async_transport.rs
@@ -147,11 +147,11 @@ where
     /// [`AsyncSmtpTransport::starttls_relay`](#method.starttls_relay) instead,
     /// if possible.
     pub fn builder_dangerous<T: Into<String>>(server: T) -> AsyncSmtpTransportBuilder {
-        let new = SmtpInfo {
+        let info = SmtpInfo {
             server: server.into(),
             ..Default::default()
         };
-        AsyncSmtpTransportBuilder { info: new }
+        AsyncSmtpTransportBuilder { info }
     }
 }
 
@@ -237,6 +237,7 @@ impl AsyncSmtpTransportBuilder {
             info: self.info,
             marker_: PhantomData,
         };
+
         AsyncSmtpTransport { inner: client }
     }
 }
